pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-eval-platform"
version = "0.1.0"
description = "Local-first LLM evaluation platform with CLI/API workflows."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "jsonschema>=4.23.0,<5.0.0",
  "pydantic>=2.11.0,<3.0.0",
  "PyYAML>=6.0.0,<7.0.0",
  "sqlalchemy>=2.0.0,<3.0.0",
  "typer>=0.16.0,<1.0.0",
  "uvicorn>=0.35.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0,<9.0.0",
]

[project.scripts]
llm-eval = "llm_eval_platform.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
llm_eval_platform = ["web/templates/*.html", "web/static/*.jsx"]
